# doc/api.yaml
openapi: 3.0.3
servers:
  - url: "http://localhost:3000"
info:
  title: Wasa Photo
  description: |-
    Wasa Photo is a photo sharing platform. It allows users to upload photo of their best moment.

  version: 1.0.0
...
paths:
...
  /profiles/{profile_user_id}/username:
    parameters:
      - name: profile_user_id
        in: path
        description: The user Id of the user who want to change username
        schema: { $ref: "#/components/schemas/UserID" }
        required: true

    put:
      tags: ["User"]
      operationId: setMyUserName
      summary: Set my new username
      description: |-
        Set new my username
      security:
        - bearerAuth: []
      requestBody:
        description: New username to set.
        content:
          application/json:
            schema:
              description: The new username of the user
              type: object
              properties:
                username: { $ref: "#/components/schemas/Username" }
        required: true
      responses:
        "200":
          description: Set new username successful
          content:
            plain/text:
              schema: { $ref: "#/components/schemas/OK" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }

  #non mi convince
  /feed:
    parameters:
      - { $ref: "#/components/parameters/Offset" }
      - { $ref: "#/components/parameters/Limit" }
    get:
      tags: ["User"]
      operationId: getMyStream
      summary: Get the user stream
      description: |-
        Get the user stream, return a list of posts sorted chronologically. See Post for more details.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Get stream successful
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PostsList" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }
...
  /profile/{profile_user_id}/bans:
    parameters:
      - name: profile_user_id
        in: path
        description: The user Id of the banned user
        schema: { $ref: "#/components/schemas/UserID" }
        required: true

    post:
      tags: ["User"]
      operationId: banUser
      summary: Ban user
      description: |-
        Ban the user with the given user ID
      security:
        - bearerAuth: []
      responses:
        "200":
          description: |-
            Ban user successful. <br>
            Return the user of the banned user. See User for more details.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }

    delete:
      tags: ["User"]
      operationId: unbanUser
      summary: Unban user
      description: |-
        Unban the user with the given user ID
      security:
        - bearerAuth: []
      responses:
        "200":
          description: |-
            Unban user successful
          content:
            text/plain:
              schema: { $ref: "#/components/schemas/OK" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }

    get:
      tags: ["User"]
      operationId: getMyBans
      summary: Get my bans
      description: |-
        Get the bans of the user with the given user ID.
        Return a list of users. See User for more details.
      security:
        - bearerAuth: []

      parameters:
        - { $ref: "#/components/parameters/Offset" }
        - { $ref: "#/components/parameters/Limit" }

      responses:
        "200":
          description: Get bans successful
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UsersList" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /profiles/{profile_user_id}:
    parameters:
      - name: profile_user_id
        in: path
        description: The user Id of the user whose profile you want to see
        schema: { $ref: "#/components/schemas/UserID" }
        required: true

    get:
      tags: ["User"]
      operationId: getUserProfile
      summary: Get user profile
      description: |-
        Get user profile
      security:
        - bearerAuth: []

      parameters:
        - { $ref: "#/components/parameters/Offset" }
        - { $ref: "#/components/parameters/Limit" }

      responses:
        "200":
          description: |-
            Get user profile successful. Return some information about the user.
            For posts list see /users/{usersID}/posts: get.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Profile" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /profiles/{profile_user_id}/posts:
    parameters:
      - name: profile_user_id
        in: path
        description: The user Id of the owner of the posts.
        schema: { $ref: "#/components/schemas/UserID" }
        required: true

    post:
      tags: ["Post"]
      operationId: uploadPhoto
      summary: Upload photo and create a post.
      description: |-
        Upload the photo and save it in the storage server.
        The user can specify the caption of the post with a max length of 64 characters.
      security:
        - bearerAuth: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              description: |-
                The photo to upload and the caption of the post.
              type: object
              properties:
                caption:
                  description: The caption of the post.
                  type: string
                  pattern: ^[^\\]{1,64}$
                  example: "In montagna con Haru"
                  minLength: 1
                  maxLength: 64
                file:
                  description: The photo to upload.
                  type: string
                  format: binary
                  minLength: 1
                  maxLength: 1000000

      responses:
        "201":
          description: Upload photo and created post successful.
          content:
            text/plain:
              schema: { $ref: "#/components/schemas/OK" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }

    get:
      tags: ["Post"]
      operationId: getPosts
      summary: Get posts
      description: |-
        Get the posts of the user with the given user ID
      security:
        - bearerAuth: []
      parameters:
        - { $ref: "#/components/parameters/Offset" }
        - { $ref: "#/components/parameters/Limit" }

      responses:
        "200":
          description: Get posts list successful
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PostsList" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }
...
  /profiles/{profile_user_id}/posts/{post_id}/like:
    parameters:
      - name: profile_user_id
        in: path
        description: The user Id of the owner of the post.
        schema: { $ref: "#/components/schemas/UserID" }
        required: true
      - name: post_id
        in: path
        description: The post id of the post to (un)like
        schema: { $ref: "#/components/schemas/PostID" }
        required: true

    post:
      tags: ["Post"]
      operationId: likePhoto
      summary: Like the post and add the user to the likes list of the post.
      description: |-
        Like post
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Liked successful
          content:
            text/plain:
              schema: { $ref: "#/components/schemas/OK" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }

    delete:
      tags: ["Post"]
      operationId: unlikePhoto
      summary: Unlike post
      description: |-
        Unlike the post and remove the user from the likes list of the post.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Unliked successful
          content:
            text/plain:
              schema: { $ref: "#/components/schemas/OK" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /profiles/{profile_user_id}/posts/{post_id}/likes:
    parameters:
      - name: profile_user_id
        in: path
        description: The user Id of the owner of the post.
        schema: { $ref: "#/components/schemas/UserID" }
        required: true
      - name: post_id
        in: path
        description: The post id of the post to see the list of users who liked it
        schema: { $ref: "#/components/schemas/PostID" }
        required: true

      - { $ref: "#/components/parameters/Offset" }
      - { $ref: "#/components/parameters/Limit" }

    get:
      tags: ["Post"]
      operationId: getLikes
      summary: Get the list of users who liked the post
      description: |-
        Get the list of users who liked the post, 
        the users object is composed by the user_id and the username.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Get the list of users who liked the post
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UsersList" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }
...
components:
  schemas:
    UserID:
      title: User id
      description: Numeric id of the user
      type: integer
      example: 10
      minimum: 1
    Username:
      title: Username
      description: Username of the user, not accepted upper case characters
      type: string
      pattern: ^[a-z][a-z0-9_]{2,13}$
      example: "haru_22"
      minLength: 3
      maxLength: 13
    User:
      title: User
      description: User
      type: object
      properties:
        user_id: { $ref: "#/components/schemas/UserID" }
        username: { $ref: "#/components/schemas/Username" }
        propic_url: { $ref: "#/components/schemas/ImgURL" }
    Profile:
      title: Profile
      description: Profile of the user
      type: object
      properties:
        user: { $ref: "#/components/schemas/User" }
        bio: { $ref: "#/components/schemas/Bio" }
        profilePic: { $ref: "#/components/schemas/ImgURL" }
        followers:
          type: integer
          description: Number of followers
          example: 10
          minimum: 0
        followings:
          type: integer
          description: Number of followings
          example: 10
          minimum: 0
        #non ritorna posts, i posts verrano ritornati con un'altra chiamata
        #forse si potrebbe usare link nei responses.
    PostID:
      title: PostID
      description: |-
        The post id is used to identify a post.
        The post id is generated by the server and it's for each user.
      type: integer
      example: 100
      minimum: 1
...
    Post:
      title: Post
      description: The post is a collection of post, like count and comment cou. Contain a timestamp.
      type: object
      properties:
        photo_url: { $ref: "#/components/schemas/ImgURL" }
        caption:
          type: string
          description: Caption of the post.
          example: "Foto al lago con Haru"
          pattern: '^[^\/\\]{1,100}$'
          minLength: 0
          maxLength: 100
        likes_count:
          type: integer
          description: Number of likes of the post.
          minimum: 0
        comments_count:
          type: integer
          description: Number of comments of the post.
          minimum: 0
        timestamp: { $ref: "#/components/schemas/Timestamp" }
    PostsList:
      title: Posts List
      description: List of posts
      type: array
      items: { $ref: "#/components/schemas/Post" }
      minItems: 0
      maxItems: 100
    UsersList:
      title: Users List
      description: |-
        A list of users for many responses.
        It's used to get a list of users for the search, follower and followings lists, banned users list and likes list.
      type: array
      items:
        type: object
        description: User
        properties:
          user: { $ref: "#/components/schemas/User" }
          proPic: { $ref: "#/components/schemas/ImgURL" }
      minItems: 0
      maxItems: 100
    Bio:
      title: Bio
      description: Bio of the user
      type: object
      properties:
        text:
          type: string
          example: "Io sono Haru, ho 1 anno e sono un gatto 🐱"
          pattern: '^[^\/\\]{1,64}$'
          minLength: 1
          maxLength: 64
...
    OK:
      title: OK
      description: The succesfull response from server.
      type: string
      example: "OK, action successfully completed"
      minLength: 2
      maxLength: 50
    ImgURL:
      title: ImgURL
      description: Image URL specification.
      type: string
      format: url
      #vorrei usare l'estensione webp
      pattern: '^https:\/\/[a-zA-Z.-_\/]*\/users\/[0-9]{5}\/(propic|posts)\/[a-zA-Z]{3}\.jpg$'
      minLength: 34
      maxLength: 128
      example: "https://www.example.com/users/00001/posts/aAk.jpg"
    Timestamp:
      title: Timestamp
      description: The timestamp of the post.
      type: string
      format: date-time
      pattern: "^[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}$"
      example: "2020-12-31T23:59:59"
      minLength: 19
      maxLength: 19
  responses:
...
    Unauthorized:
      description: The access token is missing or it's expired.
      content:
        text/plain:
          schema:
            description: Unauthorized
            type: string
            example: "Unauthorized"
            minLength: 12
            maxLength: 30
    #400
    BadRequest:
      description: The request was not compliant with the documentation (eg. missing fields, etc).
      content:
        text/plain:
          schema:
            description: BadRequest
            type: string
            example: "Bad Request"
            minLength: 11
            maxLength: 30
    #500
    InternalServerError:
      description: The server encountered an internal error. Further info in server logs.
      content:
        text/plain:
          schema:
            description: InternalServerError
            type: string
            example: "Internal Server Error"
            minLength: 21
            maxLength: 30
    #404
    NotFound:
      description: The resource was not found.
      content:
        text/plain:
          schema:
            description: NotFound
            type: string
            example: "Not Found"
            minLength: 9
            maxLength: 30
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    Offset:
      name: offset
      in: query
      description: |-
        The start index of the elements to get. <br>
        <details> 
        <summary> Other details </summary>
        If the value is **not specified**, it will be set to 0. <br>
        If the value is **greater than** the number of elements, an empty array will be returned. <br>
        If the value is **negative**, it will be set to 0. <br>
        If the value is **not an integer**, it will be set to 0.
        </details>
      schema: { type: integer, default: 0, minimum: 0 }
      required: false
    Limit:
      name: limit
      in: query
      description: |-
        The maximum number of elements to get. <br>
        <details> 
        <summary> Other details </summary>
        If the value is **not specified**, it will be set to 10. <br>          
        If the value is **greater than 100**, it will be set to 100. <br>
        If the value is **less than 1**, it will be set to 10. <br>
        If the value is **not an integer**, it will be set to 10. <br>
        If the value is **greater than** the number of elements, it will be set to the total number of elements. 
        </details>
      schema: { type: integer, default: 10, minimum: 1, maximum: 100 }
      required: false